requirements.txt
async-timeout==5.0.1
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
spotipy==2.25.1
urllib3==1.26.20
